se agrega funcion para editar y eliminar eventos del calendario con Modal

+++ src/components/Calendario/MyBigCalendar.tsx
@@ -3,17 +3,28 @@ import React, { useEffect, useState } from "react";
 import { Calendar, dayjsLocalizer } from "react-big-calendar";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import dayjs from "dayjs";
-import { CiCalendarDate } from "react-icons/ci";
+import { CiCalendarDate, CiEdit, CiEraser } from "react-icons/ci";
 import "dayjs/locale/es";
 import { getIdentity } from "../../services/user_services";
-import { getDatesByUser, deleteDate } from "../../services/date_services";
+import {
+    getDatesByUser,
+    deleteDate,
+    updateDate,
+} from "../../services/date_services";
 import { DateEntity } from "../../models/date_model";
-import { alertaError, alertaSuccess } from "../../utils/alertas";
+import {
+    alertaError,
+    alertaQuestion,
+    alertaSuccess,
+} from "../../utils/alertas";
+import EventoForm from "./EventoForm";
 
 dayjs.locale("es");
 
 export default function MyBigCalendar() {
     const [datesEntity, setDatesEntity] = useState<DateEntity[]>([]);
+    const [showModal, setShowModal] = useState(false);
+    const [editEntity, setEditEntity] = useState<DateEntity>({} as DateEntity);
 
     const localizer = dayjsLocalizer(dayjs);
     const identity = getIdentity();
@@ -22,6 +33,7 @@ export default function MyBigCalendar() {
         try {
             var tmp = await getDatesByUser(identity._id);
             setDatesEntity(tmp);
+            console.log("Fechas obtenidas", datesEntity);
         } catch (error) {
             console.log("Error al obtener fechas", error);
         }
@@ -29,13 +41,15 @@ export default function MyBigCalendar() {
 
     const deleteEvent = async (prop) => {
         try {
-            console.log("Eliminar evento", JSON.stringify(prop));
-            var deleted: boolean = await deleteDate(prop["event"]._id);
-            if (deleted) {
-                await getDates();
-                alertaSuccess("Evento eliminado");
-            } else {
-                alertaError("Error al eliminar evento");
+            var confirmar = await alertaQuestion("¿Desea eliminar el evento?");
+            if (confirmar) {
+                var deleted: boolean = await deleteDate(prop["event"]._id);
+                if (deleted) {
+                    await getDates();
+                    alertaSuccess("Evento eliminado");
+                } else {
+                    alertaError("Error al eliminar evento");
+                }
             }
         } catch (error) {
             alertaError(
@@ -44,23 +58,43 @@ export default function MyBigCalendar() {
         }
     };
 
+    const btnEditOnClic = (prop) => {
+        setEditEntity(prop);
+        setShowModal(true);
+    };
+
+    const editarEvento = async (prop: DateEntity) => {
+        prop.start = dayjs(prop.start).toDate();
+        prop.end = dayjs(prop.end).add(17, "hour").toDate();
+        var tmp: DateEntity;
+        tmp = JSON.parse(JSON.stringify(prop));
+        var updated: boolean = await updateDate(tmp);
+        if (updated) {
+            await getDates();
+            alertaSuccess("Evento actualizado");
+            setShowModal(false);
+        } else {
+            alertaError("Error al actualizar evento");
+        }
+    };
+
     const components = {
         event: (props) => {
             return (
-                <div className="flex-row gap-10 items-center">
+                <div className="flex flex-row items-center">
                     <CiCalendarDate />
                     <div>{props.title}</div>
                     <div>{props.start}</div>
                     <div>{props.end}</div>
-                    <div>
-                        <button
-                            onClick={() => console.log("Editar evento", props)}
-                        >
+                    <div className="flex">
+                        <CiEdit onClick={() => btnEditOnClic(props["event"])} />
+                        {/* <button onClick={() => btnEditOnClic(props["event"])}>
                             Editar
-                        </button>
-                        <button onClick={() => deleteEvent(props)}>
+                        </button> */}
+                        <CiEraser onClick={() => deleteEvent(props)} />
+                        {/* <button onClick={() => deleteEvent(props)}>
                             Eliminar
-                        </button>
+                        </button> */}
                     </div>
                 </div>
             );
@@ -86,6 +120,7 @@ export default function MyBigCalendar() {
                 formats={{
                     dayHeaderFormat: (date) => dayjs(date).format("DD/MM/YYYY"),
                 }}
+                views={["month", "agenda"]}
                 components={components}
                 messages={{
                     next: "Siguiente",
@@ -99,7 +134,50 @@ export default function MyBigCalendar() {
                     event: "Evento",
                     showMore: (total) => `+ Ver más (${total})`,
                 }}
+                // onSelectEvent={(slotInfo) => {
+                //     btnEditOnClic(slotInfo);
+                // }}
+                // onSelectSlot={(slotInfo) => {
+                //     console.log("Slot Seleccionado", slotInfo);
+                // }}
             />
+            {/* Modal */}
+            {showModal ? (
+                <>
+                    <div className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none">
+                        <div className="relative w-auto my-6 mx-auto max-w-3xl">
+                            {/*content*/}
+                            <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
+                                {/*header*/}
+                                <div className="flex items-start justify-between p-5 border-b border-solid border-blueGray-200 rounded-t">
+                                    <h3 className="text-3xl font-semibold">
+                                        Editar
+                                    </h3>
+                                    <button
+                                        className="p-1 ml-auto bg-transparent border-0 text-black opacity-35 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
+                                        onClick={() => setShowModal(false)}
+                                    >
+                                        <span className="bg-transparent text-black h-6 w-6 text-2xl block outline-none focus:outline-none">
+                                            X
+                                        </span>
+                                    </button>
+                                </div>
+                                {/*body*/}
+                                <div className="relative p-6 flex-auto">
+                                    <EventoForm
+                                        modelo={editEntity}
+                                        onSubmit={editarEvento}
+                                        onCancelar={() => setShowModal(false)}
+                                    />
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
+                </>
+            ) : (
+                <></>
+            )}
         </div>
     );
 }
